refactoring saved_theme

+++ lib/modules/style_manager/themes/saved_theme.dart
@@ -17,6 +17,6 @@ class SavedTheme {
 
   getPreferences() async {
     _initThemeIndex = (await _preferences.getTheme())!;
-    return _initThemeIndex as int;
+    return _initThemeIndex;
   }
 }
